+++ src/main/java/com/tidy/tidy/infrastructure/python/dto/IssueElement.java
@@ -6,10 +6,10 @@ import lombok.Getter;
 public class IssueElement {
     private Integer shapeId;
     private Integer elementIndex;
-    private Integer bboxLeft;
-    private Integer bboxTop;
-    private Integer bboxWidth;
-    private Integer bboxHeight;
+    private double bboxLeft;
+    private double bboxTop;
+    private double bboxWidth;
+    private double bboxHeight;
     private String text;
     private String elementType;
 }
